test-requirements.txt
pytest !=3.7.0, !=3.7.1  # https://github.com/python-trio/pytest-trio/pull/50#issuecomment-413124393
pytest-cov
hypothesis>=3.64
trio-asyncio
